debug rendering

+++ AVATAR/core/src/com/haruham/game/entity/HitCircle.java
@@ -9,9 +9,9 @@ import com.badlogic.gdx.math.Vector3;
 public class HitCircle {
 
     private Vector3 center;
-    private double radius;
+    private float radius;
 
-    public HitCircle(Vector3 center, double radius) {
+    public HitCircle(Vector3 center, float radius) {
         this.center = center;
         this.radius = radius;
     }
@@ -20,11 +20,11 @@ public class HitCircle {
         return center;
     }
 
-    public double getRadius() {
+    public float getRadius() {
         return radius;
     }
 
-    public void setRadius(double radius) {
+    public void setRadius(float radius) {
         this.radius = radius;
     }
